+++ components/utils.js
@@ -1,4 +1,5 @@
 import { encode } from "universal-base64";
+import yaml from "js-yaml";
 const host = process.env.NEXT_PUBLIC_HOST;
 
 export const getTicketImg = (user) => {
@@ -22,3 +23,35 @@ export const getTicketGraphImg = (user) => {
   const ticket = getTicketImg(user);
   return `https://res.cloudinary.com/demo/image/fetch/w_1200,h_630,c_pad,b_rgb:6955CB/${ticket}`;
 };
+
+function compare(a, b) {
+  // Use toUpperCase() to ignore character casing
+  const bandA = a.name.toUpperCase();
+  const bandB = b.name.toUpperCase();
+
+  let comparison = 0;
+  if (bandA > bandB) {
+    comparison = 1;
+  } else if (bandA < bandB) {
+    comparison = -1;
+  }
+  return comparison;
+}
+
+export async function getData() {
+  const context = require.context("../tracks/", false, /\.yml$/);
+  const tracks = [];
+  let speakers = [];
+  for (const key of context.keys()) {
+    const trackName = key.slice(2);
+    const content = await import(`../tracks/${trackName}`);
+    const track = yaml.safeLoad(content.default);
+    tracks.push(track);
+    const trackSpeakers = track.sessions
+      .map((s) => s.speaker)
+      .filter((sp) => !speakers.some((s) => s.name === sp.name));
+    speakers.push(...trackSpeakers);
+  }
+
+  return { tracks, speakers: speakers.sort(compare) };
+}
