+++ pages/index.js
@@ -8,7 +8,7 @@ import {
   Registration,
   Sponsors,
 } from "../components";
-import { getData } from "./api/index";
+import { getData } from "../components/utils";
 
 const Index = ({ tracks, speakers }) => {
   return (
